refactor: moving common interface to another folder
